requirements.txt
Flask==3.0.2
resend==0.7.2
python-dotenv==1.0.1 